+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Fixed
 - Index file
+### Changed
+- Rename `Checkbox` to `CheckboxInput` and `Select` to `SelectInput`  
 
 ## 0.1.3 - 2023-09-10